+++ clients/fides-js/src/lib/initialize.ts
@@ -22,23 +22,26 @@ import {
 import {
   ConsentMechanism,
   ConsentMethod,
+  CookieIdentity,
+  CookieKeyConsent,
+  CookieMeta,
   EmptyExperience,
   FidesConfig,
-  FidesOptionsOverrides,
+  FidesCookie,
   FidesOptions,
+  FidesOverrides,
+  OverrideExperienceTranslations,
   OverrideOptions,
+  OverrideType,
   PrivacyExperience,
   SaveConsentPreference,
   UserGeolocation,
-  FidesCookie,
-  CookieMeta,
-  CookieIdentity,
-  CookieKeyConsent,
 } from "./consent-types";
 import {
   constructFidesRegionString,
   debugLog,
   experienceIsValid,
+  getOverrideValidatorMapByType,
   getWindowObjFromPath,
   isPrivacyExperience,
   validateOptions,
@@ -50,7 +53,6 @@ import { updateConsentPreferences } from "./preferences";
 import { resolveConsentValue } from "./consent-value";
 import { initOverlay } from "./consent";
 import { TcfCookieConsent } from "./tcf/types";
-import { FIDES_OVERRIDE_OPTIONS_VALIDATOR_MAP } from "./consent-constants";
 import { setupExtensions } from "./extensions";
 import {
   noticeHasConsentInCookie,
@@ -198,51 +200,55 @@ const automaticallyApplyGPCPreferences = async ({
 };
 
 /**
- * Gets and validates override options provided through URL query params, cookie, or window obj
+ * Gets and validates overrides provided through URL query params, cookie, or window obj
+ * This shared fn supports getting 2 different types of overrides: FidesOptionsOverrides and
+ * FidesExperienceLanguageOverrides
  *
- * If the same override option is provided in multiple ways, load the value in this order:
+ * If the same override is provided in multiple ways, load the value in this order:
  * 1) query param  (top priority)
  * 2) window obj   (second priority)
  * 3) cookie value (last priority)
  */
-export const getOptionsOverrides = (
+export const getOverridesByType = <T>(
+  type: OverrideType,
   config: FidesConfig
-): Partial<FidesOptionsOverrides> => {
-  const overrideOptions: Partial<FidesOptionsOverrides> = {};
+): Partial<T> => {
+  const overrides: Partial<T> = {};
   if (typeof window !== "undefined") {
     // Grab query params if provided in the URL (e.g. "?fides_string=123...")
     const queryParams = new URLSearchParams(window.location.search);
-    // Grab override options if exists (e.g. window.fides_overrides = { fides_string: "123..." })
+    // Grab window overrides if exists (e.g. window.fides_overrides = { fides_string: "123..." })
     const customPathArr: "" | null | string[] =
       config.options.customOptionsPath &&
       config.options.customOptionsPath.split(".");
-    const windowObj: OverrideOptions | undefined =
+    const windowObj:
+      | Partial<OverrideOptions & OverrideExperienceTranslations>
+      | undefined =
       customPathArr && customPathArr.length >= 0
         ? getWindowObjFromPath(customPathArr)
         : window.fides_overrides;
 
     // Look for each of the override options in all three locations: query params, window object, cookie
-    FIDES_OVERRIDE_OPTIONS_VALIDATOR_MAP.forEach(
-      ({ fidesOption, fidesOptionType, fidesOverrideKey, validationRegex }) => {
-        const queryParamOverride: string | null =
-          queryParams.get(fidesOverrideKey);
+    const overrideValidatorMap = getOverrideValidatorMapByType(type);
+    overrideValidatorMap?.forEach(
+      ({ overrideName, overrideType, overrideKey, validationRegex }) => {
+        const queryParamOverride: string | null = queryParams.get(overrideKey);
         const windowObjOverride: string | boolean | undefined = windowObj
-          ? windowObj[fidesOverrideKey]
+          ? windowObj[overrideKey]
           : undefined;
-        const cookieOverride: string | undefined =
-          getCookieByName(fidesOverrideKey);
+        const cookieOverride: string | undefined = getCookieByName(overrideKey);
 
-        // Load the override option value, respecting the order of precedence (query params > window object > cookie)
+        // Load the override value, respecting the order of precedence (query params > window object > cookie)
         const value = queryParamOverride || windowObjOverride || cookieOverride;
         if (value && validationRegex.test(value.toString())) {
-          // coerce to expected type in FidesOptions
-          overrideOptions[fidesOption] =
-            fidesOptionType === "string" ? value : JSON.parse(value.toString());
+          // coerce to expected type
+          overrides[overrideName as keyof T] =
+            overrideType === "string" ? value : JSON.parse(value.toString());
         }
       }
     );
   }
-  return overrideOptions;
+  return overrides;
 };
 
 /**
@@ -328,6 +334,7 @@ export const initialize = async ({
   geolocation,
   renderOverlay,
   updateExperience,
+  overrides,
 }: {
   cookie: FidesCookie;
   savedConsent: CookieKeyConsent;
@@ -347,6 +354,7 @@ export const initialize = async ({
     debug?: boolean;
     isExperienceClientSideFetched: boolean;
   }) => Partial<PrivacyExperience>;
+  overrides?: Partial<FidesOverrides>;
 } & FidesConfig): Promise<Partial<Fides>> => {
   let shouldInitOverlay: boolean = options.isOverlayEnabled;
   let effectiveExperience = experience;
@@ -438,7 +446,13 @@ export const initialize = async ({
       if (shouldInitOverlay) {
         // Initialize the i18n singleton before we render the overlay
         const i18n = setupI18n();
-        initializeI18n(i18n, window?.navigator, effectiveExperience, options);
+        initializeI18n(
+          i18n,
+          window?.navigator,
+          effectiveExperience,
+          options,
+          overrides?.experienceTranslationOverrides
+        );
 
         // OK, we're (finally) ready to initialize & render the overlay!
         await initOverlay({