+++ clients/privacy-center/pages/api/fides-js.ts
@@ -171,6 +171,7 @@ export default async function handler(
       preventDismissal: environment.settings.PREVENT_DISMISSAL,
       allowHTMLDescription: environment.settings.ALLOW_HTML_DESCRIPTION,
       base64Cookie: environment.settings.BASE_64_COOKIE,
+      fidesPrimaryColor: environment.settings.FIDES_PRIMARY_COLOR,
     },
     experience: experience || undefined,
     geolocation: geolocation || undefined,